Blackoffer/requirements.txt
Django==4.0.4
django_filter==22.1
djangorestframework==3.14.0
